# Optimize `BlackjackEnv` `_get_obs`