--- a/gymnasium/envs/toy_text/blackjack.py
+++ b/gymnasium/envs/toy_text/blackjack.py
@@ -45,6 +45,18 @@ def is_natural(hand):  # Is this hand a natural blackjack?
     return sorted(hand) == [1, 10]
 
 
+def _hand_sum_and_usable_ace(hand):
+    """
+    Helper to compute both sum with and without ace counted as 11,
+    and whether the hand has a usable ace.
+    Returns (effective_sum, usable_ace: int)
+    """
+    s = sum(hand)
+    if 1 in hand and s + 10 <= 21:
+        return s + 10, 1
+    return s, 0
+
+
 class BlackjackEnv(gym.Env):
     """
     Blackjack is a card game where the goal is to beat the dealer by obtaining cards
@@ -196,7 +208,9 @@ class BlackjackEnv(gym.Env):
         return self._get_obs(), reward, terminated, False, {}
 
     def _get_obs(self):
-        return (sum_hand(self.player), self.dealer[0], usable_ace(self.player))
+        # Optimization: Compute sum and usable ace in one pass per step, not two.
+        player_sum, player_usable_ace = _hand_sum_and_usable_ace(self.player)
+        return (player_sum, self.dealer[0], player_usable_ace)
 
     def reset(
         self,