# Description

<!-- CODEFLASH_OPTIMIZATION: {"function":"BlackjackEnv._get_obs","file":"gymnasium/envs/toy_text/blackjack.py","speedup_pct":"60%","speedup_x":"0.60x","original_runtime":"163 microseconds","best_runtime":"102 microseconds","optimization_type":"general","timestamp":"2025-08-07T04:14:52.785Z","version":"1.0"} -->
### 📄 60% (0.60x) speedup for ***`BlackjackEnv._get_obs` in `gymnasium/envs/toy_text/blackjack.py`***

⏱️ Runtime :   **`163 microseconds`**  **→** **`102 microseconds`** (best of `251` runs)
### 📝 Explanation and details


The optimization eliminates redundant `sum(hand)` calculations that were happening multiple times in the original code. 

**Key changes:**
- **Introduced `_hand_sum_and_usable_ace` helper function** that computes both the hand sum and usable ace status in a single pass, avoiding duplicate `sum(hand)` calls
- **Modified `_get_obs`** to call the helper once instead of separately calling `sum_hand()` and `usable_ace()`, which each performed their own sum calculations
- **Preserved original function signatures** for `usable_ace` and `sum_hand` while making them use the shared computation internally

**Why this is faster:**
In the original code, `_get_obs()` called both `sum_hand()` and `usable_ace()`. The `sum_hand()` function internally called `usable_ace()`, creating a cascade where `sum(hand)` was computed 3 times per observation:
1. Once in the direct `usable_ace()` call
2. Once in `sum_hand()` when it calls `usable_ace()` 
3. Once more in `sum_hand()` for the final sum calculation

The optimization reduces this to a single `sum(hand)` calculation per observation, explaining the **60% speedup**.

**Best performance gains** are seen in test cases with:
- **Large hands** (e.g., 999 cards showing 171% speedup) where `sum()` becomes expensive
- **Hands with multiple aces** (40-50% speedups) where the ace logic is more complex
- **Any scenario calling `_get_obs`** frequently, as this is the primary bottleneck eliminated

The line profiler confirms this: original `_get_obs` took 1.691ms total, while optimized version takes only 0.443ms - a direct result of eliminating redundant sum calculations.



✅ **Correctness verification report:**

| Test                        | Status            |
| --------------------------- | ----------------- |
| ⚙️ Existing Unit Tests | 🔘 **None Found** |
| 🌀 Generated Regression Tests | ✅ **127 Passed** |
| ⏪ Replay Tests | 🔘 **None Found** |
| 🔎 Concolic Coverage Tests | 🔘 **None Found** |
|📊 Tests Coverage       | 100.0% |
<details>
<summary>🌀 Generated Regression Tests and Runtime</summary>

```python
import gymnasium as gym
# imports
import pytest  # used for our unit tests
from gymnasium import spaces
from gymnasium.envs.toy_text.blackjack import BlackjackEnv

# unit tests

# --------------------------
# Basic Test Cases
# --------------------------

def test_basic_no_ace():
    # Player has 10 and 7, dealer shows 5
    env = BlackjackEnv()
    env.player = [10, 7]
    env.dealer = [5, 9]
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.81μs -> 1.50μs (20.2% faster)

def test_basic_with_usable_ace():
    # Player has Ace and 7, dealer shows 10
    env = BlackjackEnv()
    env.player = [1, 7]
    env.dealer = [10, 3]
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.01μs -> 1.48μs (35.9% faster)

def test_basic_with_nonusable_ace():
    # Player has Ace, 10, 5 (Ace can't be 11)
    env = BlackjackEnv()
    env.player = [1, 10, 5]
    env.dealer = [8, 2]
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.98μs -> 1.47μs (34.1% faster)

def test_basic_dealer_ace():
    # Dealer shows Ace, player has 2,3
    env = BlackjackEnv()
    env.player = [2, 3]
    env.dealer = [1, 9]
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.80μs -> 1.48μs (21.9% faster)

def test_basic_dealer_ten():
    # Dealer shows 10, player has 2,3
    env = BlackjackEnv()
    env.player = [2, 3]
    env.dealer = [10, 8]
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.84μs -> 1.47μs (25.1% faster)

# --------------------------
# Edge Test Cases
# --------------------------

def test_edge_minimum_player_sum():
    # Player has [2,2], dealer shows 2 (lowest possible non-bust hand)
    env = BlackjackEnv()
    env.player = [2, 2]
    env.dealer = [2, 7]
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.84μs -> 1.44μs (28.3% faster)

def test_edge_maximum_player_sum():
    # Player has [10,10,1] (should be 21, ace as 1)
    env = BlackjackEnv()
    env.player = [10, 10, 1]
    env.dealer = [10, 5]
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.18μs -> 1.56μs (40.0% faster)

def test_edge_maximum_player_sum_with_usable_ace():
    # Player has [1,10] (natural blackjack)
    env = BlackjackEnv()
    env.player = [1, 10]
    env.dealer = [10, 4]
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.95μs -> 1.43μs (36.8% faster)

def test_edge_multiple_aces_one_usable():
    # Player has [1, 1, 8] (one ace usable)
    env = BlackjackEnv()
    env.player = [1, 1, 8]
    env.dealer = [9, 7]
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.01μs -> 1.47μs (36.6% faster)

def test_edge_multiple_aces_none_usable():
    # Player has [1, 1, 10, 9] (all aces must be 1)
    env = BlackjackEnv()
    env.player = [1, 1, 10, 9]
    env.dealer = [7, 2]
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.06μs -> 1.42μs (45.2% faster)

def test_edge_dealer_card_1():
    # Dealer shows 1 (Ace), player has [5, 6]
    env = BlackjackEnv()
    env.player = [5, 6]
    env.dealer = [1, 10]
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.85μs -> 1.44μs (29.2% faster)

def test_edge_dealer_card_10():
    # Dealer shows 10, player has [1, 8]
    env = BlackjackEnv()
    env.player = [1, 8]
    env.dealer = [10, 3]
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.00μs -> 1.41μs (41.9% faster)

def test_edge_bust_hand():
    # Player has [10, 10, 5] (bust)
    env = BlackjackEnv()
    env.player = [10, 10, 5]
    env.dealer = [4, 4]
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.83μs -> 1.51μs (20.8% faster)

def test_edge_empty_hand():
    # Player has no cards, dealer has [2,3]
    env = BlackjackEnv()
    env.player = []
    env.dealer = [2, 3]
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.61μs -> 1.28μs (25.5% faster)

def test_edge_single_card_hand():
    # Player has [1], dealer has [3,4]
    env = BlackjackEnv()
    env.player = [1]
    env.dealer = [3, 4]
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.97μs -> 1.43μs (38.1% faster)

    # Player has [10], dealer has [7,8]
    env.player = [10]
    env.dealer = [7, 8]
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.07μs -> 785ns (36.4% faster)

# --------------------------
# Large Scale Test Cases
# --------------------------

def test_large_hand_no_ace():
    # Player has 50 cards of 2 (total 100), dealer shows 5
    env = BlackjackEnv()
    env.player = [2] * 50
    env.dealer = [5, 7]
    codeflash_output = env._get_obs(); obs = codeflash_output # 3.17μs -> 2.35μs (35.3% faster)

def test_large_hand_with_aces_some_usable():
    # Player has 10 aces, 5 twos (should count as as many aces as possible as 11, but not bust)
    env = BlackjackEnv()
    env.player = [1]*10 + [2]*5
    env.dealer = [9, 8]
    # Let's compute expected sum and usable_ace:
    # Try to use as many aces as 11 as possible without busting.
    # Start with all aces as 1: 10 + 10 = 20
    # Add 5*2 = 10 --> 20
    # Now, try to promote aces to 11: each promotion adds +10, but must not exceed 21.
    # Only one ace can be promoted: 20 + 10 = 30 (bust), so none usable.
    # Actually, sum([1]*10 + [2]*5) = 10 + 10 = 20, but that's not correct: 10 + 10 = 20.
    # Wait, sum([1]*10) = 10, sum([2]*5) = 10, total = 20.
    # If we make one ace usable: 20 + 10 = 30 (bust). So no usable ace.
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.29μs -> 1.53μs (49.3% faster)

def test_large_hand_with_usable_ace():
    # Player has [1, 2, 2, 2, 2, 2], dealer shows 7
    env = BlackjackEnv()
    env.player = [1] + [2]*5  # sum = 1+10=11, with ace as 11: 21
    env.dealer = [7, 3]
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.12μs -> 1.45μs (46.1% faster)

def test_large_dealer_hand():
    # Dealer has 100 cards, player has [3,4]
    env = BlackjackEnv()
    env.player = [3, 4]
    env.dealer = [2]*100
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.78μs -> 1.45μs (22.5% faster)

def test_large_varied_hand():
    # Player has 100 cards: 50 aces, 50 tens
    env = BlackjackEnv()
    env.player = [1]*50 + [10]*50
    env.dealer = [10, 10]
    # sum = 50 + 500 = 550
    # Try to make as many aces as 11 as possible, but each ace promoted adds +10, but will bust
    # All aces must be 1, so sum_hand=550, usable_ace=0
    codeflash_output = env._get_obs(); obs = codeflash_output # 4.33μs -> 2.25μs (92.4% faster)

def test_large_many_aces_one_usable():
    # Player has 1 ace and 9 ones (total 10 cards), dealer shows 10
    env = BlackjackEnv()
    env.player = [1] + [1]*9
    env.dealer = [10, 10]
    # sum = 10, with ace as 11: 20, usable_ace=1
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.15μs -> 1.50μs (43.3% faster)

def test_large_all_aces_bust():
    # Player has 22 aces (should be 22, all as 1)
    env = BlackjackEnv()
    env.player = [1]*22
    env.dealer = [1, 2]
    # sum = 22, can't have usable ace
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.43μs -> 1.54μs (57.8% faster)

def test_large_maximum_hand():
    # Player has 999 cards of 1 (aces), dealer shows 5
    env = BlackjackEnv()
    env.player = [1]*999
    env.dealer = [5, 6]
    # sum = 999, no usable ace
    codeflash_output = env._get_obs(); obs = codeflash_output # 23.0μs -> 8.49μs (171% faster)

def test_large_mixed_hand_with_usable_ace():
    # Player has 1 ace, 9 twos, dealer shows 10
    env = BlackjackEnv()
    env.player = [1] + [2]*9
    env.dealer = [10, 8]
    # sum = 1+18=19, with ace as 11: 29 (bust), so ace is not usable
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.19μs -> 1.53μs (42.7% faster)

def test_large_mixed_hand_with_usable_ace2():
    # Player has 1 ace, 5 twos, dealer shows 10
    env = BlackjackEnv()
    env.player = [1] + [2]*5  # sum = 1+10=11, with ace as 11: 21
    env.dealer = [10, 3]
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.10μs -> 1.43μs (46.8% faster)
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.
#------------------------------------------------
import gymnasium as gym
# imports
import pytest  # used for our unit tests
from gymnasium import spaces
from gymnasium.envs.toy_text.blackjack import BlackjackEnv

# -------------------------------
# UNIT TESTS FOR _get_obs FUNCTION
# -------------------------------

# Helper function to create env and set hands
def setup_env(player_hand, dealer_hand):
    env = BlackjackEnv()
    env.player = player_hand
    env.dealer = dealer_hand
    return env

# 1. BASIC TEST CASES

def test_basic_no_ace():
    # Player: 7, 8; Dealer: 10
    env = setup_env([7, 8], [10, 6])
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.82μs -> 1.45μs (25.2% faster)

def test_basic_with_usable_ace():
    # Player: Ace, 7; Dealer: 6
    env = setup_env([1, 7], [6, 5])
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.02μs -> 1.50μs (34.8% faster)

def test_basic_with_nonusable_ace():
    # Player: Ace, 9, 5; Dealer: 3
    env = setup_env([1, 9, 5], [3, 2])
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.03μs -> 1.48μs (37.2% faster)

def test_basic_blackjack():
    # Player: Ace, 10; Dealer: 7
    env = setup_env([1, 10], [7, 4])
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.03μs -> 1.45μs (40.2% faster)

def test_basic_dealer_ace():
    # Player: 5, 6; Dealer: Ace
    env = setup_env([5, 6], [1, 7])
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.81μs -> 1.46μs (23.7% faster)

# 2. EDGE TEST CASES

def test_edge_player_bust():
    # Player: 10, 9, 5; Dealer: 2
    env = setup_env([10, 9, 5], [2, 8])
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.86μs -> 1.48μs (25.8% faster)

def test_edge_player_min_sum():
    # Player: 2, 2; Dealer: 3
    env = setup_env([2, 2], [3, 4])
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.81μs -> 1.42μs (27.7% faster)

def test_edge_player_max_sum_no_bust():
    # Player: 10, 10, Ace; Dealer: 8
    env = setup_env([10, 10, 1], [8, 9])
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.21μs -> 1.59μs (39.2% faster)

def test_edge_multiple_aces_one_usable():
    # Player: Ace, Ace, 8; Dealer: 5
    env = setup_env([1, 1, 8], [5, 7])
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.01μs -> 1.45μs (38.0% faster)

def test_edge_multiple_aces_none_usable():
    # Player: Ace, Ace, 10, 10; Dealer: 4
    env = setup_env([1, 1, 10, 10], [4, 2])
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.07μs -> 1.49μs (39.0% faster)

def test_edge_player_sum_21_with_multiple_aces():
    # Player: Ace, Ace, 9; Dealer: 10
    env = setup_env([1, 1, 9], [10, 2])
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.03μs -> 1.42μs (43.3% faster)

def test_edge_dealer_showing_ace():
    # Player: 10, 2; Dealer: Ace
    env = setup_env([10, 2], [1, 8])
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.86μs -> 1.45μs (28.6% faster)

def test_edge_dealer_showing_ten():
    # Player: 8, 7; Dealer: 10
    env = setup_env([8, 7], [10, 3])
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.85μs -> 1.45μs (27.6% faster)

def test_edge_player_all_aces():
    # Player: Ace, Ace, Ace; Dealer: 2
    env = setup_env([1, 1, 1], [2, 3])
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.01μs -> 1.42μs (41.6% faster)

def test_edge_player_all_tens():
    # Player: 10, 10, 10; Dealer: 5
    env = setup_env([10, 10, 10], [5, 8])
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.86μs -> 1.45μs (28.3% faster)

def test_edge_player_empty_hand():
    # Player: []; Dealer: 7
    env = setup_env([], [7, 4])
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.66μs -> 1.27μs (30.8% faster)

def test_edge_dealer_empty_hand():
    # Player: 5, 6; Dealer: []
    env = setup_env([5, 6], [])
    # Should raise IndexError when accessing dealer[0]
    with pytest.raises(IndexError):
        env._get_obs() # 2.78μs -> 2.44μs (14.1% faster)

def test_edge_player_single_card():
    # Player: 10; Dealer: 3
    env = setup_env([10], [3, 2])
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.75μs -> 1.41μs (24.0% faster)

def test_edge_player_single_ace():
    # Player: Ace; Dealer: 4
    env = setup_env([1], [4, 5])
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.98μs -> 1.34μs (47.9% faster)

# 3. LARGE SCALE TEST CASES

def test_large_player_hand_no_ace():
    # Player: 100 cards of 2; Dealer: 10
    player_hand = [2] * 100
    dealer_hand = [10, 7]
    env = setup_env(player_hand, dealer_hand)
    codeflash_output = env._get_obs(); obs = codeflash_output # 4.66μs -> 3.23μs (44.1% faster)

def test_large_player_hand_with_aces_some_usable():
    # Player: 10 Aces, 9 cards of 1 (Ace), 10 cards of 2
    player_hand = [1]*9 + [2]*10
    dealer_hand = [5, 9]
    env = setup_env(player_hand, dealer_hand)
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.39μs -> 1.58μs (51.4% faster)

def test_large_player_hand_with_usable_ace():
    # Player: Ace, 2, 3, 4, 5, 6, 7, 8, 9, 10 (10 cards)
    player_hand = [1,2,3,4,5,6,7,8,9,10]
    dealer_hand = [3, 7]
    env = setup_env(player_hand, dealer_hand)
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.15μs -> 1.49μs (44.6% faster)

def test_large_dealer_hand():
    # Player: 7, 8; Dealer: 100 cards of 2
    player_hand = [7, 8]
    dealer_hand = [2]*100
    env = setup_env(player_hand, dealer_hand)
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.83μs -> 1.45μs (26.0% faster)

def test_large_player_all_aces():
    # Player: 20 Aces; Dealer: 10
    player_hand = [1]*20
    dealer_hand = [10, 7]
    env = setup_env(player_hand, dealer_hand)
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.44μs -> 1.57μs (55.7% faster)

def test_large_player_sum_just_under_21_with_usable_ace():
    # Player: Ace, 9
    player_hand = [1, 9]
    dealer_hand = [8, 2]
    env = setup_env(player_hand, dealer_hand)
    codeflash_output = env._get_obs(); obs = codeflash_output # 1.98μs -> 1.41μs (40.2% faster)

def test_large_player_sum_exactly_21_with_usable_ace():
    # Player: Ace, 9, Ace
    player_hand = [1, 9, 1]
    dealer_hand = [8, 2]
    env = setup_env(player_hand, dealer_hand)
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.02μs -> 1.45μs (39.1% faster)

def test_large_player_sum_just_over_21_with_aces():
    # Player: Ace, 9, 2
    player_hand = [1, 9, 2]
    dealer_hand = [8, 2]
    env = setup_env(player_hand, dealer_hand)
    codeflash_output = env._get_obs(); obs = codeflash_output # 2.01μs -> 1.44μs (39.3% faster)

def test_large_dealer_showing_various():
    # Dealer showing 1-10, player: 10, 7
    for i in range(1, 11):
        env = setup_env([10, 7], [i, 3])
        codeflash_output = env._get_obs(); obs = codeflash_output # 8.13μs -> 5.54μs (46.8% faster)

def test_large_player_hand_max_size():
    # Player: 999 cards of 1 (ace), Dealer: 10
    player_hand = [1]*999
    dealer_hand = [10, 5]
    env = setup_env(player_hand, dealer_hand)
    codeflash_output = env._get_obs(); obs = codeflash_output # 22.9μs -> 8.44μs (172% faster)
# codeflash_output is used to check that the output of the original code is the same as that of the optimized code.
```

</details>


To edit these changes `git checkout codeflash/optimize-BlackjackEnv._get_obs-me0vv8ji` and push.

[![Codeflash](https://img.shields.io/badge/Optimized%20with-Codeflash-yellow?style=flat&color=%23ffc428&logo=data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iNDgwIiBoZWlnaHQ9ImF1dG8iIHZpZXdCb3g9IjAgMCA0ODAgMjgwIiBmaWxsPSJub25lIiB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciPgo8cGF0aCBmaWxsLXJ1bGU9ImV2ZW5vZGQiIGNsaXAtcnVsZT0iZXZlbm9kZCIgZD0iTTI4Ni43IDAuMzc4NDE4SDIwMS43NTFMNTAuOTAxIDE0OC45MTFIMTM1Ljg1MUwwLjk2MDkzOCAyODEuOTk5SDk1LjQzNTJMMjgyLjMyNCA4OS45NjE2SDE5Ni4zNDVMMjg2LjcgMC4zNzg0MThaIiBmaWxsPSIjRkZDMDQzIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMzExLjYwNyAwLjM3ODkwNkwyNTguNTc4IDU0Ljk1MjZIMzc5LjU2N0w0MzIuMzM5IDAuMzc4OTA2SDMxMS42MDdaIiBmaWxsPSIjMEIwQTBBIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMzA5LjU0NyA4OS45NjAxTDI1Ni41MTggMTQ0LjI3NkgzNzcuNTA2TDQzMC4wMjEgODkuNzAyNkgzMDkuNTQ3Vjg5Ljk2MDFaIiBmaWxsPSIjMEIwQTBBIi8+CjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xpcC1ydWxlPSJldmVub2RkIiBkPSJNMjQyLjg3MyAxNjQuNjZMMTg5Ljg0NCAyMTkuMjM0SDMxMC44MzNMMzYzLjM0NyAxNjQuNjZIMjQyLjg3M1oiIGZpbGw9IiMwQjBBMEEiLz4KPC9zdmc+Cg==)](https://codeflash.ai)

## Type of change

- [x] New feature (non-breaking change which adds functionality)

# Checklist:

…